Makes AutoConfigure and MiniClusterUtil use proper HBase config
This fixes https://github.com/Mayo-AAU/HadoopCommons/issues/5.

Testing can now be done anywhere.  The HBase startup uses the
hostname/domain as a bind address.  Over VPN and sometimes on WiFi,
this does not resolve correctly because of DNS update issues.  The fix
is to set the hostname explicitly to `localhost` is in:

```
# Don't lookup the hostname, specify localhost
hbase.master.hostname=localhost
hbase.regionserver.hostname=localhost
```